backend/requirements.backend.txt
# --- Server
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
sse-starlette==2.1.3

# --- LangChain + LangServe (ecosistema 0.3.x, compatible con langserve 0.3.x)
langserve==0.3.3
langchain==0.3.27
langchain-community==0.3.31
langchain-text-splitters==0.3.11
langchain-openai==0.3.11

# --- Vector store (Chroma)
langchain-chroma==0.2.6
chromadb>=1.0.20


# --- Loaders / scraping helpers
beautifulsoup4==4.12.3
lxml==5.3.0
tqdm==4.67.1
